+++ packages/ui/src/react/dialog.tsx
@@ -155,7 +155,10 @@ export function DialogHeader({
 }: React.ComponentProps<'header'>): React.JSX.Element {
   return (
     <header
-      className={twMerge('grid shrink-0 gap-2 border-b px-6 py-3', className)}
+      className={twMerge(
+        'py-3.75 grid shrink-0 gap-2 border-b px-6',
+        className,
+      )}
       {...props}
     />
   );
@@ -185,7 +188,7 @@ export function DialogFooter({
   return (
     <footer
       className={twMerge(
-        'flex shrink-0 flex-col-reverse gap-2 border-t px-6 py-3 sm:flex-row sm:justify-between',
+        'py-3.75 flex shrink-0 flex-col-reverse gap-2 border-t px-6 sm:flex-row sm:justify-between',
         className,
       )}
       {...props}